# api-docs-mon-archive.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/missions/{code}": {
      "get": {
        "tags": [
          "mission-controller"
        ],
        "operationId": "findByCode",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MissionResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "mission-controller"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MissionResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "mission-controller"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/dossiers/{codeDossier}": {
      "get": {
        "tags": [
          "dossier-controller"
        ],
        "operationId": "getByCode",
        "parameters": [
          {
            "name": "codeDossier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DossierResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "dossier-controller"
        ],
        "operationId": "update_1",
        "parameters": [
          {
            "name": "codeDossier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DossierUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DossierResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "dossier-controller"
        ],
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "codeDossier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/dossiers": {
      "get": {
        "tags": [
          "dossier-controller"
        ],
        "operationId": "getAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DossierResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "dossier-controller"
        ],
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DossierCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DossierResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/upload": {
      "post": {
        "tags": [
          "document-controller"
        ],
        "operationId": "upload",
        "parameters": [
          {
            "name": "codeDossier",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "codeMission",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents": {
      "get": {
        "tags": [
          "document-controller"
        ],
        "operationId": "list",
        "parameters": [
          {
            "name": "codeDossier",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "codeMission",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/download/{id}": {
      "get": {
        "tags": [
          "document-controller"
        ],
        "operationId": "download",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit/utilisateur": {
      "get": {
        "tags": [
          "audit-controller"
        ],
        "operationId": "getByUtilisateur",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLog"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/audit/ressource": {
      "get": {
        "tags": [
          "audit-controller"
        ],
        "operationId": "getByRessource",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLog"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit/periode": {
      "get": {
        "tags": [
          "audit-controller"
        ],
        "operationId": "getByPeriode",
        "parameters": [
          {
            "name": "debut",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "fin",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLog"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit/action": {
      "get": {
        "tags": [
          "audit-controller"
        ],
        "operationId": "getByAction",
        "parameters": [
          {
            "name": "action",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "CREER_DOSSIER",
                "CREER_MISSION",
                "CONSTITUER_EQUIPE",
                "CONSULTER_DOCUMENT",
                "TELECHARGER_DOCUMENT",
                "SUPPRIMER_DOCUMENT",
                "SUPPRIMER_DOSSIER",
                "UPDATE_DOSSIER",
                "UPLOAD_DOCUMENT",
                "SUPPRIMER_MISSION",
                "UPDATE_MISSION"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLog"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents/{id}": {
      "delete": {
        "tags": [
          "document-controller"
        ],
        "operationId": "delete_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "MissionUpdateRequest": {
        "type": "object",
        "properties": {
          "statut": {
            "type": "string",
            "enum": [
              "PLANIFIEE",
              "EN_COURS",
              "TERMINEE",
              "SUSPENDUE"
            ]
          },
          "exercice": {
            "type": "string"
          }
        }
      },
      "MissionResponse": {
        "type": "object",
        "properties": {
          "codeMission": {
            "type": "string"
          },
          "exercice": {
            "type": "string"
          },
          "codeDossier": {
            "type": "string"
          },
          "statut": {
            "type": "string",
            "enum": [
              "PLANIFIEE",
              "EN_COURS",
              "TERMINEE",
              "SUSPENDUE"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DossierUpdateRequest": {
        "required": [
          "raisonSociale"
        ],
        "type": "object",
        "properties": {
          "raisonSociale": {
            "type": "string"
          },
          "typeEntreprise": {
            "type": "string"
          },
          "adresse": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "bp": {
            "type": "string"
          }
        }
      },
      "DossierResponse": {
        "type": "object",
        "properties": {
          "codeDossier": {
            "type": "string"
          },
          "raisonSociale": {
            "type": "string"
          },
          "typeEntreprise": {
            "type": "string"
          },
          "adresse": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "bp": {
            "type": "string"
          }
        }
      },
...
      "DossierCreateRequest": {
        "required": [
          "codeDossier",
          "raisonSociale"
        ],
        "type": "object",
        "properties": {
          "codeDossier": {
            "type": "string"
          },
          "raisonSociale": {
            "type": "string"
          },
          "typeEntreprise": {
            "type": "string"
          },
          "adresse": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "bp": {
            "type": "string"
          }
        }
      },
      "DocumentResponse": {
        "type": "object",
        "properties": {
          "nomFichierOriginal": {
            "type": "string"
          },
          "nomFichierStocke": {
            "type": "string"
          },
          "cheminStockage": {
            "type": "string"
          },
          "typeFichier": {
            "type": "string"
          },
          "taille": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "downloadUrl": {
            "type": "string"
          }
        }
      },
...
      "AuditLog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "action": {
            "type": "string",
            "enum": [
              "CREER_DOSSIER",
              "CREER_MISSION",
              "CONSTITUER_EQUIPE",
              "CONSULTER_DOCUMENT",
              "TELECHARGER_DOCUMENT",
              "SUPPRIMER_DOCUMENT",
              "SUPPRIMER_DOSSIER",
              "UPDATE_DOSSIER",
              "UPLOAD_DOCUMENT",
              "SUPPRIMER_MISSION",
              "UPDATE_MISSION"
            ]
          },
          "performedBy": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "performedAt": {
            "type": "string",
            "format": "date-time"
          },
          "entityType": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "ipAddress": {
            "type": "string"
          }
        }
      }
    }
  }
}